ECL: Remove unit-test time overheads.
